merge: merge

+++ srcs/b_export.c
@@ -6,7 +6,7 @@
 /*   By: naykim <naykim@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/13 16:34:17 by naykim            #+#    #+#             */
-/*   Updated: 2022/01/20 16:16:38 by naykim           ###   ########.fr       */
+/*   Updated: 2022/01/24 16:04:17 by naykim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void	ft_export(t_var *var, char *new)
 		return ;
 	new_key = ft_substr(new, 0, ft_strchr(new, '=') - new);
 	i = -1;
-	while (++i < ft_sstrlen(var->our_env))
+	while (++i < (int)ft_sstrlen(var->our_env))
 	{
 		if (!ft_strncmp(var->our_env[i], new_key, ft_strlen(new_key)))
 		{
@@ -44,7 +44,7 @@ void	b_export(t_var *var, char **cmd)
 	if (!(var->our_env) || !cmd)
 		return ;
 	i = 1;
-	while (i < ft_sstrlen(cmd))
+	while (i < (int)ft_sstrlen(cmd))
 	{
 		ft_export(var, cmd[i]);
 		i++;
